files moved to coincide with src
